Move profile into separate module + little tweaks
